epuka app
Epuka is a system intended to help the recovery people cope up with the issue they are undergoing through. The system is implemented on a android platform, which has the capacity of reaching out to hundreds of people as well as sek help from a large pool of people and professionals willing to offer there help.